Add cos, sin and tests
Add cos and sin constexpr functions and corresponding
method for dual numbers.

Add tests for exp, cos and sin, checking that the result is the same
as std:: algorithms (note: equal up to numerical precision, for larger
values than the ones tested it might be needed to increase the
number of elements of the taylor series).