Define helpers for geometric series and sums
Define `impl::series::{geometric, sum}` to reduce boilerplate when
implementing functions as sums of geometric series. `geometric_series`
implements an input range view that generates the terms in a geometric
series given an initial count, initial value, and a function providing
the multiplicative factor for the next term in the series.